refactor, test-fix: deal with new sanity check findings
