+++ test/functionality/dataflow/dataflow.spec.ts
@@ -2,11 +2,13 @@ import { requireAllTestsInFolder } from '../_helper/collect-tests'
 import path from 'path'
 
 describe('Dataflow', () => {
-	require('./environments/environments')
+	describe('Environments', () =>
+		requireAllTestsInFolder(path.join(__dirname, 'environments'))
+	)
 
-	describe('Graph', () => {
+	describe('Graph', () =>
 		requireAllTestsInFolder(path.join(__dirname, 'graph'))
-	})
+	)
 
 
 	require('./processing-of-elements/processing-of-elements')
